Edits according to Sam's comments
