+++ src/layouts/index.tsx
@@ -88,7 +88,6 @@ export default class Layout extends React.Component<LayoutProps, LayoutState> {
             maxWidth: 960,
             padding: '0px 1.0875rem 1.45rem',
             paddingTop: 0,
-            marginBottom: 130,
           })}
         >
           {this.props.children()}
@@ -98,6 +97,7 @@ export default class Layout extends React.Component<LayoutProps, LayoutState> {
             margin: '0 auto',
             maxWidth: 960,
             padding: '0px 1.0875rem 1.45rem',
+            marginBottom: 130,
           })}
         >
           <label>
@@ -132,9 +132,7 @@ export default class Layout extends React.Component<LayoutProps, LayoutState> {
               bottom: 17,
               position: 'fixed',
             })}
-            scrollamount="20"
             direction="left"
-            behavior="scroll"
           >
             <img
               className={css({
